@@ -7,26 +7,33 @@
   "license": "MIT",
   "scripts": {
     "dev": "next",
+    "build": "next build",
+    "start": "next start",
     "test": "mocha --compilers js:babel-core/register --require ./test/setup.js './**/*Specs.js'",
     "test:watch": "npm run test -- --watch"
   },
   "dependencies": {
     "auth0-lock": "^10.9.0",
-    "eslint-config-airbnb": "latest",
+    "glamor": "^2.20.24",
+    "isomorphic-fetch": "^2.2.1",
     "js-cookie": "^2.1.3",
     "jwt-decode": "^2.1.0",
     "moment": "^2.17.1",
-    "next": "^1.2.3",
+    "next": "2.0.0-beta",
+    "next-redux-wrapper": "^1.0.0",
     "ramda": "^0.23.0",
     "react-redux": "^5.0.1",
     "redux": "^3.6.0",
-    "redux-thunk": "^2.1.0",
+    "redux-devtools-extension": "^2.13.0",
+    "redux-thunk": "^2.2.0",
     "uuid": "^3.0.1"
   },
   "devDependencies": {
     "babel-eslint": "^7.1.1",
+    "babel-preset-stage-0": "^6.22.0",
     "chai": "^3.5.0",
     "eslint": "^3.13.0",
+    "eslint-config-airbnb": "^14.1.0",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^3.0.2",
     "eslint-plugin-react": "^6.9.0",
