@@ -3,7 +3,8 @@
  *
  * @author pkapako
  */
-const reset = `
+export default () => (
+    <style jsx global>{`
 html, body, div, span, applet, object, iframe,
 h1, h2, h3, h4, h5, h6, p, blockquote, pre,
 a, abbr, acronym, address, big, cite, code,
@@ -47,5 +48,5 @@ table {
   border-collapse: collapse;
   border-spacing: 0;
 }
-`;
-export default reset;
+`}</style>
+);
